TLC-552 QA6 add run command and refactor
