+++ qa6/helpers/userDelete.js
